requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-browser-reload==1.13.0
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-tailwind==3.8.0
idna==3.7
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
